+++ varlisp/fmt_print_visitor.hpp
@@ -39,9 +39,9 @@ struct fmt_print_visitor : public boost::static_visitor<void> {
         // or will call m_fmt.print<bool>()
         m_fmt.print(m_o, sss::string_view("nil"));
     }
-    void operator()(bool   v) const { m_fmt.print(m_o, v); }
-    void operator()(int    v) const { m_fmt.print(m_o, v); }
-    void operator()(double v) const { m_fmt.print(m_o, v); }
+    void operator()(bool    v) const { m_fmt.print(m_o, v); }
+    void operator()(int64_t v) const { m_fmt.print(m_o, v); }
+    void operator()(double  v) const { m_fmt.print(m_o, v); }
     void operator()(const string_t&             v) const;
     void operator()(const varlisp::symbol&      s) const;
     void operator()(const varlisp::List&        l) const;
