+++ varlisp/is_instant_visitor.hpp
@@ -25,19 +25,19 @@ struct is_instant_visitor : public boost::static_visitor<bool> {
         return false;
     }
 
-    bool operator()(const varlisp::symbol& ) const { return false; }
-    bool operator()(const varlisp::List& v) const;
+    bool operator()(const varlisp::symbol&    ) const { return false; }
+    bool operator()(const varlisp::List& v    ) const ;
 
-    bool operator()(const varlisp::Lambda& ) const { return true; }
+    bool operator()(const varlisp::Lambda&    ) const { return true; }
     bool operator()(const sss::regex::CRegex& ) const { return true; }
-    bool operator()(const gumboNode& ) const { return true; }
-    bool operator()(const Empty&) const { return true; }
-    bool operator()(const Nill&) const { return true; }
-    bool operator()(bool ) const { return true; }
-    bool operator()(int ) const { return true; }
-    bool operator()(double ) const { return true; }
-    bool operator()(const std::string ) const { return true; }
-    bool operator()(const varlisp::Builtin& ) const { return true; }
+    bool operator()(const gumboNode&          ) const { return true; }
+    bool operator()(const Empty&              ) const { return true; }
+    bool operator()(const Nill&               ) const { return true; }
+    bool operator()(bool                      ) const { return true; }
+    bool operator()(int64_t                   ) const { return true; }
+    bool operator()(double                    ) const { return true; }
+    bool operator()(const std::string         ) const { return true; }
+    bool operator()(const varlisp::Builtin&   ) const { return true; }
 };
 }  // namespace varlisp
